[BUGFIX] Allow poll author to see votings, when finishing poll
When "Super Secret Mode" poll setting is enabled, the poll author still
needs to see the amount of votes, to be able to select a final option.

In finish view, the poll author always see the results.

Also, when the poll is finished, everyone see the votes and results (without participants).