[!!!][TASK] Refactor Permission Event handling
The old slot-adapted solution had a high potential for attacking itself
and thus getting stuck in an event caller loop.

The modification made here places the event handling on the
PollController and allows for finer adjustment, as each step is now
stored individually as an event. On the one hand, this facilitates
troubleshooting and, on the other hand, prevents an unintentional
event-in-event loop, which was previously the case.

At the same time, exception handling is optimised and case-specific
exceptions are thrown, which clearly identify the permissions as the
source of the error and which permission is actually missing at that
point.

The controller itself has been modified so that all actions now return a
ResponseInterface, which will suffice for a later upgrade to TYPO3 v12.