[TASK] Make response changeable by UpdateAfterEvent
Before refactoring to EventListeners, the Slots could create redirects
after manipulating the data.
This technique could be used again, but for getting a clearer state, it
is better to use response manipulation through the event and the
respective Listeners ensuring the response being returned directly
inside the controller instead of some other places.

Create the default redirect response as variable inside the controller
action and override with the manipulated one from the event.

This ensures a clean state and proper possibilities for the manipulation